Added EventEmitter example implementations.
Two examples were created which is how we'd imagine this functionality would be utilized.
a) Subcribe to multiple events via the on mehod
b) The use of method chaining and the magic call methods. The example shows how the EventEmitter inmplementation (in this case MockEventEmitter but this is also Restler class itself) can be accessed statically without having a reference, from anywhece it the application.