Added support for Resources.json recognizing HTTPS when behind Amazon Elastic Load Balancer
AWS ELB can present an HTTPS interface to itself but communicate to its backend servers via HTTP. When this occurs, the client's browser thinks all traffic is via HTTPS, but the backend server thinks all traffic is via HTTP. To provide a method for backend servers to recognize the front-end protocol, Amazon provides a header "HTTP_X_FORWARDED_PROTO".

This commit recognizes the Amazon header and correctly sets Restler::$baseUrl to an https:// address when traffic is recieved via HTTP but the client is communicating to an HTTPS load balancer.