Add support for access to raw request data
Incoming data was being parsed automatically by Restler library. There was no option to access raw data for example for doing a data checksum, hash its content for check that the query is legitime and so on. The pull consists in defining a new attribute which contains  the raw data, which is initialized while reading from php://input.